Keep MainPing for backwards compatibility
